Add egs_phsp_scoring ausgab object.
This ausgab object allows the user to score phase space files in EGSnrc or
IAEA format for particles entering/exiting a predefined geometry.
See the documentation at the the top of egs_phsp_scoring.cpp for more
details.

It has been tested for a single scoring plane perpendicular to the Z axis.
However, testing for other and more complex geometries is encouraged and
required!

The commit also includes a small change to the check_byte_order() function
in iaea_header.cpp: The function, which is called every time an IAEA header
file is updated, no longer outputs the byte order (along with an outdated
list of associated architectures) to the screen.  This was resulting in
a lot of cryptic screen output when writing IAEA phase space files.