Fix a tutor7pp ausgab bug
Fix a tutor7pp bug related to ausgab calls. The value of whether or not
to do russian roulette was not initialized properly. When an ausgab
object was used in the input file, results differed due so random memory
accesses. The results were correct when no ausgab objects were used.

This is therefore a critical bug fix for tutor7pp.

Also add some ausgab call values that were missing in application.h.
This change is just for future use and consistency.