Fix predefined data for SSB media in pegsless mode.
Precalculated density effect data available for a small
subset of media, referred to in pegs4 as
Sternheimer-Seltzer-Berger (SSB) media, were not included
in the pegsless implementation, even though the
lookup table for the names of these media was.  This
caused a crash in the (rare) pegsless run in which an
SSB medium was referred to by its exact name
(e.g. "H2O" instead of "h2o" or "H2OICRU512," etc) and
a density correction file was not specified.  The
precalculated data for these media is now included
in the pegsless implementation, although really just to
be consistent with pegs4.