Fix the tutor7pp PHD regions
Fix a bug in tutor7pp where the pulse height regions specified by the
user in the input file were not associated with the correct regions. The
PHDs that were output corresponded to ir+1, but the input/output
suggested that they corresponded to ir. Now, the PHDs are output for the
correct regions and there is additional description of the output
format.