Fix #388: concave cd geometry errors
Fix a special case where convex geometries result in a concave cd
geometries when combined. This showed up as geometry errors and could be
created using a radius-mismatched cylinder and sphere in a cd geometry.
Also fix where a small step is taken in cylinder or sphere howfar
routines to step by boundaryTolerance/2 instead of boundaryTolerance.
This resulted in checks against the boundaryTolerance to fail in the cd
geometry.