Try and fix egs++ egs_mesh CI builds by enforcing C++11
The current g++ is 5.4, around the time C++11 was introduced.
I believe the default feature set for 5.4 is C++03. Testing a local build
with -std=c++03 caused the egs++ build to fail like CI.

Enabling -std=c++11 instead led to a successful build of egs++.
The mevegs user code will still be broken until we link against the Gmsh SDK.