Add mevegs application based on tutor7pp
We'd like to write an output file based on the mesh when the
simulation is done. This is most convenient to do in something
like Mevegs_Application::outputResults but requires specific
knowledge of the EGS_Mesh geometry (e.g. region volumes). We
dynamic_cast the base geometry to a EGS_Mesh pointer to access
its data and prepare to save a Gmsh file.